ai_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
langchain==0.2.6
langgraph==0.1.19
google-generativeai>=0.7.0
supabase>=2.10.0
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
